pyproject.toml
[tool.poetry]
name = "options_magic_data_collection"
version = "0.1.0"
description = ""
authors = ["<user> <user> <<user>@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
yfinance = "^0.1.63"
pandas = "^1.3.3"
numpy = "^1.21.2"
yahoo-fin = "^0.8.9.1"
requests-html = "^0.10.0"
psycopg2-binary = "^2.9.10"
jsonschema = "^4.23.0"
html5lib = "^1.1"
pytz = "^2025.1"
python-dotenv = "^1.0.1"
scipy = "^1.15.2"
supabase = "^2.27.2"
pillow = "^12.1.1"
tweepy = "^4.16.0"
selenium = "^4.41.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"


[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
